+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Other blocks are:
 - **API**: Added support of resonance Raman activity as `QLabel(quantity="ramact", descriptor="RR")`.
 - **API**: Support of electronic transition property derivatives from the formatted checkpoint file.
 - **API**: The index of the excited electronic state for vibronic calculations can be extracted from Gaussian log files.
+- **API**: `DataError` class for general errors related to operations on extracted data.
 
 ### Fixed
 - **ESP**: Failure to plot J matrix on recent versions of Python.